requirements.txt
feedparser==6.0.1
SQLAlchemy==1.3.19
pycoingecko==1.3.0
alembic==1.4.3
python-dotenv==0.14.0
mysqlclient==2.0.1
pycoingecko==2.2.0